+++ blocks/analog_constant.h
@@ -1,7 +1,7 @@
 struct analog_constant_options {
-	int value;
+	int16_t value;
 };
 
-void analog_constant (struct analog_constant_options *options, int *out0) {
+void analog_constant (struct analog_constant_options *options, int16_t *out0) {
 	*out0 = options->value;
 }
