+++ blocks/analog_input.h
@@ -2,6 +2,6 @@ struct analog_input_options {
 	int channel;
 };
 
-void analog_input (struct analog_input_options *options, int *out0) {
+void analog_input (struct analog_input_options *options, int16_t *out0) {
 	*out0 = read_adc(options->channel);
 }
