+++ blocks/analog_logging.h
@@ -2,4 +2,4 @@ struct analog_logging_options {
 	int channel;
 };
 
-void analog_logging (struct analog_logging_options *options, int *in0) { }
+void analog_logging (struct analog_logging_options *options, int16_t *in0) { }
