+++ blocks/saturation.h
@@ -1,9 +1,9 @@
 struct saturation_options {
-	int high_limit;
-	int low_limit;
+	int16_t high_limit;
+	int16_t low_limit;
 };
 
-void saturation(struct saturation_options *options, int *in0, int *out0) {
+void saturation(struct saturation_options *options, int16_t *in0, int16_t *out0) {
 	if (*in0 > options->high_limit) {
 		*out0 = options->high_limit;
 	} else if (*in0 < options->low_limit) {
